// ClientApp/generated/JobExecutionStatusModel.ts
import { JobResultModel } from './JobResultModel';
/*  GENERTAED automatically using Models.tst. DO NOT CHANGE HERE */
import { jobStatusModel } from './jobStatusModel';
//Generated from class commonInterfaces.dbDataTypes.JobExecutionStatusModel
export interface JobExecutionStatusModel {
  status: jobStatusModel;
  result: JobResultModel;
}