// ClientApp/generated/jobStatusModel.ts
/*  GENERTAED automatically using Models.tst. DO NOT CHANGE HERE */

//Generated from class commonInterfaces.dbDataTypes.jobStatusModel
export enum jobStatusModel {
  created = 'created',
  checking = 'checking',
  fetched = 'fetched',
  inprogress = 'inprogress',
  completed = 'completed',
  failed = 'failed',
}